[BOOKINGSG-9134][Nghi] rename the dropdown list v2 and update file import
