+++ src/input-group/types.ts
@@ -6,7 +6,7 @@ import type {
     DropdownSearchProps,
     ItemsLoadStateType,
     TruncateType,
-} from "../shared/dropdown-list-v2";
+} from "../shared/dropdown-list";
 
 export type InputGroupAddonType = "label" | "list" | "custom";
 export type InputGroupAddonPosition = "left" | "right" | undefined;
