+++ src/input-nested-multi-select/types.ts
@@ -6,7 +6,7 @@ import type {
 import type {
     DropdownCustomLabelProps,
     DropdownSearchProps,
-} from "../shared/dropdown-list-v2/types";
+} from "../shared/dropdown-list/types";
 
 // =============================================================================
 // INPUT SELECT PROPS
