+++ src/input-nested-select/input-nested-select.tsx
@@ -6,13 +6,13 @@ import { useEffect, useRef, useState } from "react";
 import type {
     NestedDropdownListItemProps,
     NestedDropdownListLocalItem,
-} from "../shared/dropdown-list-v2";
+} from "../shared/dropdown-list";
 import {
     buildKeyPathToSet,
     ExpandableElement,
     findItemByKeyPath,
     NestedDropdownList,
-} from "../shared/dropdown-list-v2";
+} from "../shared/dropdown-list";
 import { ElementWithDropdown } from "../shared/dropdown-wrapper";
 import {
     LabelContainer,
