+++ src/input-nested-select/types.ts
@@ -9,7 +9,7 @@ import type {
     DropdownSearchProps,
     DropdownVariantType,
     ExpandMode,
-} from "../shared/dropdown-list-v2/types";
+} from "../shared/dropdown-list/types";
 import type { DropdownAlignmentType } from "../shared/dropdown-wrapper/types";
 
 // =============================================================================
