+++ src/input-range-select/types.ts
@@ -8,7 +8,7 @@ import type {
     DropdownDisplayProps,
     DropdownSearchProps,
     ItemsLoadStateType,
-} from "../shared/dropdown-list-v2/types";
+} from "../shared/dropdown-list/types";
 import type { DropdownAlignmentType } from "../shared/dropdown-wrapper";
 
 export interface InputRangeSelectOptionsProps<T>
