+++ src/input-select/input-select.tsx
@@ -6,7 +6,7 @@ import {
     DropdownList,
     DropdownListState,
     ExpandableElement,
-} from "../shared/dropdown-list-v2";
+} from "../shared/dropdown-list";
 import { ElementWithDropdown } from "../shared/dropdown-wrapper";
 import {
     LabelContainer,
