+++ src/select-histogram/types.ts
@@ -1,7 +1,7 @@
 import type { AriaAttributes, RefObject } from "react";
 
 import type { HistogramSliderProps } from "../histogram-slider";
-import type { TruncateType } from "../shared/dropdown-list-v2/types";
+import type { TruncateType } from "../shared/dropdown-list/types";
 import type { DropdownAlignmentType } from "../shared/dropdown-wrapper";
 
 export type SelectHistogramSliderProps = Pick<
