+++ src/shared/dropdown-list/expandable-element.tsx
