+++ src/shared/dropdown-wrapper/dropdown-wrapper.styles.tsx
@@ -1,10 +1,7 @@
 import styled, { css } from "styled-components";
 
 import { V3_Colour, V3_Font } from "../../v3_theme";
-import type {
-    DropdownVariantType,
-    TruncateType,
-} from "../dropdown-list-v2/types";
+import type { DropdownVariantType, TruncateType } from "../dropdown-list/types";
 import { lineClampCss } from "../styles";
 
 // =============================================================================
