+++ src/time-range-picker/combobox-picker/combobox-picker.tsx
@@ -8,9 +8,9 @@ import {
 } from "../../form/form-label.style";
 import { ClearIcon } from "../../input/input.style";
 import { ClearIconContainer } from "../../input-range-select/input-range-select.style";
-import { concatIds,VisuallyHidden } from "../../shared/accessibility";
-import { DropdownListState } from "../../shared/dropdown-list-v2";
-import { DropdownList } from "../../shared/dropdown-list-v2/dropdown-list";
+import { concatIds, VisuallyHidden } from "../../shared/accessibility";
+import { DropdownListState } from "../../shared/dropdown-list";
+import { DropdownList } from "../../shared/dropdown-list/dropdown-list";
 import { ElementWithDropdown } from "../../shared/dropdown-wrapper";
 import { RangeInputInnerContainer } from "../../shared/range-input-inner-container";
 import { SimpleIdGenerator } from "../../util";
